Поддержка AVIF [deploy]

+++ tools/images.js
@@ -5,13 +5,16 @@ import { processImage } from "#server/lib/image.js";
 const src = `${cwd}/tmp/recipe`;
 const files = await readdir(src);
 
-await Promise.all(
-	files
-		.filter((file) => /jpg|png|webp$/.test(file))
-		.map(async (filename) => {
-			const filePath = `${src}/${filename}`;
-			const file = await readFile(filePath);
+for (const filename of files) {
+	if (/jpg|png|webp$/.test(filename)) {
+		const filePath = `${src}/${filename}`;
+		const file = await readFile(filePath);
+		try {
 			await processImage(file, filename.replace(/\..*$/, ""));
 			await unlink(filePath);
-		}),
-);
+			console.info(`${filename}: процесс завершён успешно!`);
+		} catch (error) {
+			console.error(error);
+		}
+	}
+}
